Remove width and height parameters from Image.
The width and height parameters are just passed through as the source
width and height of the image, not as any sort of scaling value. The
width and height should always match the image height, and since it is
known for the images used in the Go API, there's no point in passing it
in.