Outlined shapes.
Added outlined versions of the shapes because otherwise you would have
to draw them with a fill colour of zero alpha which isn't very efficient
as the fill still happens (as opposed to the stroke not happening if
strokewidth is set to zeo).

Changed the newpath() to only create VG_CAPABILITY_APPEND_TO paths
since the driver may be able to optimize and there is no usage for
the other capabilities at yet.

Changed the MAXFONTPATH to be consistent with the declaration.